Create fizzBuzz.js
A function that outputs (console.log) numbers in the range from begin to end to the terminal. At the same time:

If the number is divisible by 3, then the word Fizz is output instead.                 
If the number is divisible by 5, then the word Buzz is output instead.             
If the number is divisible by both 3 and 5, then the word FizzBuzz is output instead of the number
In other cases, the number itself is output

The function accepts two parameters (begin and end) that define the beginning and end of the range (inclusive). For simplicity, we assume that these parameters are integers greater than zero. If the range is empty (in the case when begin > end), then the function simply does not print anything.

Calling a function:
fizzBuzz(11, 20);

Terminal output:
11
Fizz
13
14
FizzBuzz
16
17
Fizz
19
Buzz